feat: Loan 생성 로직, kafka 최초 req 발행 로직
Former-commit-id: e7ef04e9c9ddeb5c663152f5c9b47ac103e48135

+++ back/contract-service/src/main/java/com/mosaic/loan/dto/CreateLoanRequestDto.java
@@ -1,10 +1,11 @@
 package com.mosaic.loan.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record CreateLoanRequestDto(
-        long id,
-        int requestAmount,
+        Integer id,
+        BigDecimal requestAmount,
         LocalDate due_date
 ) {
 }
